requirements.txt
torch~=2.2.2
sentence-transformers~=5.1.0
pypdf~=6.0.0
pandas~=2.3.2
<<<<<<< HEAD
scikit-learn
numpy~=1.26.4
joblib
=======
scikit-learn~=1.7.1
numpy~=1.26.4
joblib~=1.5.2
>>>>>>> recovered-work
h11~=0.16.0
pip~=25.2
filelock~=3.19.1
requests~=2.32.5
typing_extensions~=4.15.0
sniffio~=1.3.1
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
certifi~=2025.8.3
urllib3~=2.5.0
packaging~=25.0
pydantic~=2.11.7
pillow~=11.3.0
anyio~=4.10.0
tqdm~=4.67.1
pydantic_core~=2.33.2
annotated-types~=0.7.0
charset-normalizer~=3.4.3
Jinja2~=3.1.6
ollama~=0.5.3
protobuf~=5.29.5
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.74.0
proto-plus~=1.26.1
pyasn1_modules~=0.4.2
cachetools~=5.5.2
google-api-python-client~=2.181.0
grpcio-status~=1.71.2
httplib2~=0.30.0
pyparsing~=3.2.3
uritemplate~=4.2.0
google-auth-httplib2~=0.2.0
<<<<<<< HEAD
python-dotenv~=1.1.1
=======
python-dotenv~=1.1.1
fastapi~=0.116.1
optuna~=4.5.0
shap~=0.48.0
matplotlib~=3.10.6
lightgbm~=4.6.0
>>>>>>> recovered-work
